Populate updated = added Veterinarians to DB. Implemented veterinarians to database in a new table.

+++ src/main/java/dat/dto/VeterinarianDTO.java
@@ -8,7 +8,7 @@ import lombok.*;
 @AllArgsConstructor
 public class VeterinarianDTO {
 
-    private Long id;           // Unique identifier
+    private Integer id;           // Unique identifier
     private String name;       // Name of the veterinarian
     private String specialty;  // Veterinarian's specialty
     private String phone;      // Contact phone number
